+++ ak_library_management/controllers/__init__.py
@@ -1,1 +1,1 @@
-from . import contacts_controller
+from . import main